# forge/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GamePanel API",
    "description": "Game server management panel API with server provisioning, user management, and node orchestration.",
    "version": "0.1.0",
    "contact": {
      "name": "GamePanel Team",
      "url": "https://github.com/anomalyco/gamepanel"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/v1/account/sessions": { "get": { "summary": "List sessions", "tags": ["Account"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Session list" } } } },
    "/api/v1/account/sessions/{id}": { "delete": { "summary": "Revoke session", "tags": ["Account"], "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Session revoked" } } } },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "apiKey": { "type": "apiKey", "in": "header", "name": "Authorization" }
    },
...
  },
...
}